requirements.txt
asgiref==3.7.2
dj-database-url==2.1.0
Django==5.0.1
django-phonenumber-field==7.3.0
djangorestframework==3.14.0
drf-yasg==1.21.7
gunicorn==21.2.0
inflection==0.5.1
packaging==23.2
phonenumbers==8.13.29
psycopg2-binary==2.9.9
pytz==2023.4
PyYAML==6.0.1
sqlparse==0.4.4
typing_extensions==4.9.0
uritemplate==4.1.1
